requirements.txt
grpcio==1.38.1
grpcio-tools==1.38.1
lxml==4.7.1
numpy==1.19.5
protobuf==3.15.8
sounddevice==0.4.1
